fixes / improvements to some spr_sprani_sumo_fe_cvt_Exst previous submissions
9FC88069_1024x512 reinstated the graphic equalizer bar element
1762489B_512x128 sme pixelds from the "???" were interfering with the "average rank" text
A737D1C1_512x512 interferences betwen different textures fixed
ACF61D7C_1024x512 redone the "START" label which is actually used (overimposed to ALL courses in the shop, the one in the textures there is covered). Also, added the TM to the logo because it's there in that specific menu page in the original one.
CBF8ECBF_1024x512_text_only (png) as already discussed on Discord, the text part for this one.