Default Button Prompts and Related Textures
To be faithful to the original textures, I decided to leave the remade keyboard prompts as the default textures for button prompts.
Why? As we know, OR2006 sprite organization is a mess and sometimes default graphics (like loading, backgrounds or logos for example) are mixed with button prompts. Not everyone has a controller attached to their computer, so keyboard makes sense to be the new standard.